Save progress, connection works if referential are set to 0
Why the setting is set to 0? 
It tells to the LDAP connection to not look for help from another servers.
This basically, makes the search easier by only allowing to find the request for the server indicated on the fetch function.
So the code does not need to try to look for other servers, reducing the problems and changes of getting errors