Remove duplicated section. It's similar to the Session Cookie PHPSESSID
This section about the "remember me" its very similar to the one about the "PHPSESSID".

The PHPSESSID cookie is native to PHP and enables websites to store serialised state data.

This property is also used to identify a visitor on the portal.

This feature is not implementated on the portal, I thought it was because its a stantard